51-dracut-rescue-postinst.sh: Create hmac file for rescue kernel
When fips=1 in enabled on a server, and the server is booted into the generated rescue kernel grub menu entry. It will fail to boot since there isn't an hmac file for the rescue vmlinuz file. This patch checks if the sha512hmac binary exist, if it does it generates the hmac file for the rescue kernel. Also it checks if a rescue kernel entry exist in the grub.cfg, if it does it doesn't add a new entry. This is so when ran on existing rescue kernels, it doesn't create duplicate entries when generating the hmac. https://bugzilla.redhat.com/show_bug.cgi?id=923439

Signed-off-by: Ryan Blakley <rblakley@redhat.com>